design: header, player bar media query
