+++ src/components/playbar/Play.tsx
@@ -1,15 +1,20 @@
 import { useDispatch, useSelector } from 'react-redux'
 import { RootState } from '../../redux/config/configStore'
-import styled from 'styled-components'
 import useAudio from '../../hooks/useAudio'
 import React, { useEffect, useRef, useState, useCallback } from 'react'
 import playBar from '../../assets/icons/music_play_brown.png'
 import playStopBar from '../../assets/icons/music_stop_brown.png'
 import { setTogglePlaying } from '../../redux/modules/isPlaying'
-
-interface Props {
-  width: string
-}
+import {
+  AudioWrap,
+  ProgressBar,
+  Dealt,
+  AudioContain,
+  PlayExplain,
+  PlayTitle,
+  PlayImg,
+  PlayImgContain,
+} from './PlaySt'
 
 function Play() {
   const [handleTimeUpdate, audioRef, setMusicNumber] = useAudio()
@@ -86,9 +91,13 @@ function Play() {
             <p style={{ color: '#888888' }}>{data.musicPlayer.composer}</p>
           </PlayExplain>
           {playingState === false ? (
-            <PlayImg src={playBar} onClick={onClickPlayToggleHandler} />
+            <PlayImgContain>
+              <PlayImg src={playBar} onClick={onClickPlayToggleHandler} />
+            </PlayImgContain>
           ) : (
-            <PlayImg src={playStopBar} onClick={onClickPlayToggleHandler} />
+            <PlayImgContain>
+              <PlayImg src={playStopBar} onClick={onClickPlayToggleHandler} />
+            </PlayImgContain>
           )}
           {currentTime === 0 ||
           currentTime === undefined ||
@@ -104,63 +113,15 @@ function Play() {
             </p>
           )}
         </AudioContain>
-        <audio
-          ref={audioRef}
-          onTimeUpdate={handleTimeUpdate}
-          src={data.musicPlayer.musicUrl}
-          autoPlay={true}
-        />
       </AudioWrap>
+      <audio
+        ref={audioRef}
+        onTimeUpdate={handleTimeUpdate}
+        src={data.musicPlayer.musicUrl}
+        autoPlay={true}
+      />
     </>
   )
 }
 
 export default React.memo(Play)
-
-const AudioWrap = styled.div`
-  width: 100%;
-  height: 120px;
-  background-color: #fafafa;
-  border: 1px solid #dfdfdf;
-  box-sizing: border-box;
-  position: fixed;
-  bottom: 0px;
-  margin-top: 30px;
-`
-const ProgressBar = styled.div`
-  width: 100%;
-  height: 8px;
-  background-color: #d9d9d9;
-`
-const Dealt = styled.div<Props>`
-  background-color: #4b372e;
-  width: ${(props) => props.width + '%'};
-  height: 5px;
-  border-radius: 5px;
-`
-
-const AudioContain = styled.div`
-  width: 1280px;
-  height: 120px;
-  margin: auto;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: relative;
-`
-const PlayExplain = styled.div`
-  width: 400px;
-`
-const PlayTitle = styled.p`
-  white-space: nowrap;
-  overflow: hidden;
-  font-size: 20px;
-`
-const PlayImg = styled.img`
-  width: 70px;
-  height: 70px;
-  cursor: pointer;
-  margin: auto;
-  position: absolute;
-  left: 600px;
-`
